fix(ui5-application-inquirer): move tests
